Define initial data of Engine structure
I have chosen to pass to the engine the initial data that it needs in form of a struct. This has the benefit that we could add members to it without affecting other parts of the code and also that it reduces the number of parameters we need to pass to the constructor of the engine class and possibly to other functions that might require some of the information in this structure.