Add MovePlayer() to Physics system
Physics namespace represents our physics system whose functions take in the required data and process them. For now it is a namespace filled with free functions as there was no need to create a struct or class as it does not hold any state so a namespace of functions made more sense. MovePlayer() essentially checks if there have been any change in position for the player recorded and then makes sure what has been passed as entity is actually the player and then it changes its position.