Create Entity and its relevant member
Each entity is composed mainly of EntityData that contains all the possible data that is most likely going to be common amongst all kinds of entities we decide to create. We could extend this struct and create more specialized types of entities that could possibly have their own extra states.