Do keyboard handling outside of SDL_PollEvent()
The problem with SDL_PollEvent() is that when you hold a key it does not record an event for a short while before repeatedly recording the holding of the key event until you let go. This is problematic for when we expect continuous movement when we hold the key. This behaviour is the same as that of any OS when you hold a key. To circumvent this, we use SDL_GetKeyboardState(), which always returns an array of bools indicating the state of the key at every frame. Use this inside ProcessKeyboardInput() and write ProcessKeyboardInput() outside of SDL_PollEvent() in order to avoid the mentioned behaviour of it.