Eliminate physics system awarness of fields of engine
Physics system should be completely oblivious to what happens in the engine and so it should not have any access to any of the fields of the engine. This decoupling simplifies the physics code and makes it more useable. That is why we eliminated the queue parameter and made the engine pass on the delta pos value to the MovePlayer() free function instead of its queue.