Make PlayerMovementComponent resp. for only movement
PlayerMovementComponent was not responsible for creating the deltaTransform that is used now in its Update() method. Instead, the PlayerInputComponent was the one responsible for doing that which is against the seperation of concern principle and makes it hard to add or remove features in the future due to such coupling of responsibilities.