Sync rotation of ship with mouse
SDL3 way of rotating sprites boils down to passing the angle of clockwise rotation to one of its render texture functions. So there is no point in including the rotation part in m_transform as it wont be used. We have to remove this and just register the displacement as a 2d vector.