Make entity connector public member
Since components are going to reference the entity connector to get data from other components of entities or entities themselves, it is best for it to be public as there is no point in making it private. A getter is unnecessary since the connector needs to be non const which defeats the point of getter in here.