Do not construct EntityPool with its type
If we were to construct the EntityPool with its type then it would mean that we would have to pass all the types of the pools that the EntitySpawnerFromPools needs to consrtruct the respective pools. But this would enforce a needless order of types that need to pass to the constructor of the spawner and also it makes less sense that Init the whole pool with all the required information.