Create Entity pool
The entity pool defined here functions roughly as follows: there are 2 vectors where each is the total number of entities you create beforehand when adding entities to the pool at initialization. The only allowed values the elements of both vectors can have are either the actual index of the entity in the vector of entities or std::numeric_limits<uint32_t>::max(). The first vector in each update checks to see whether the corresponding entity is active or not. If it is active then it sets the element to uint32_t::max(), if not it sets the element to the index of the inactive entity. We then scan this array again to find the first index of an inactive entity. The second array stores from oldest active to newest active. This vector gets all its meaningful indices compressed at the beginning of the vector so that we would have a section of meaningful active indices of entities and second section of uint32_t::max(). This second vector is needed to keep track of oldest to newest in order to reuse the same active ones when the pool runs out of inactive indices.