Create class for sending requests between entities
An entity or component of an entity might need some data from another entity or component. So we create an entity connector class for sending the requests between entities and getting the data that each of needs. The implementation caches the most used entities indices or beginning indices of the entities based on their types. This was we spend less time iterating through the vector of entities.