Reference entities by handle through the engine
Using ptr to entities is dangerous since the vector of entitites could get resized which invalidates that ptr. Therefore, we should reference the entities through their index and for that we need to reference the engine in order to obtain reference to that entity every frame. That is why we intiialize components with a ptr to engine.