Pass ptr of engine and enable to set windw res used
As explained before we keep track of the owner entities through their indices through a ptr to the engine. We also get to use the current window res of the window in that frame in order to avoid rendering entities that go out of the scenes limits. We also include a setting of window res that is used by this graphics component. This is necessary since the window res can change between frame and so we need to ensure the correct res is used in each update.