Keep track of non-repeated keys and decouple registering and processing of inputs and flushing of non repeated keys array before processing events
When an event is triggered by pressed a key, due to how input is handled by OS, the key is registered several times. This is kept track by the array with repetition word in it. The problem is ensued when we temp hold the key(which happens everything we press it since CPU are fast at processing frames), this causes there to be no events to process as the while loop has processed all events related to a key being pressed. This will cause the bullets to be used from the pool in consecutive adjacent frames which causes several bullets to be shot when we press the key corresponding to shooting of bullets. To fix this we flush(memset to false) the array of non-repetition keys before processing events of SDL.

We also decoupled the registering of mouse and key from processing them since this makes the code abide by the law of seperation of concern and be easier to maintain.