Define LOGGING macro in LogSystem.h file directly
The way we should use the logging system is by defining the LOGGING macro in the LogSystem.h file directly so that we just need to include this file whenever we need to log. This way when we need to not log anymore without removing the included log statements throughout the code; we just comment out this macro in the LogSystem.h file and compile.