Create getter&&setters for entities && systems
It is important to only pass what is needed from the engine to the object that requests it. So we only pass the entities requested through their handles and not pass the whole vector of entitites managed by the engine itself. We also pass ref to other systems in const which avoids modifying them.