Clamp the speed when the ship is not being moved
When the fps drops, due to the equations used for smoothing out the movement of the ship when it stops moving, l_deltaTime increases and could possibly cause the speed to actually increase. So we need to clamp the lv_tempSpeed in order to make sure the ship comes to a stop when it is not being moved by WASD.

This bug was detected when i moved the ship and stopped moving it and continually pressed F to shoot. Due to the logging associated with shooting lasers, the fps dropped and we could see the ship moving even though it stopped being moved when we just shot bullets.