Modify components to purely based on behaviours
The components behaviour was based on the inheritance and name of the entity that they were supposed to be added to. This is quite limiting as it was pointed out in the previous reviews, since it will a lot of the times lead to repeating the same thing under different class component names and also is not modular and reusable.