Simplify array of key States that are independent
Since the action of key down for allowed repetition keys and key up are independent we can combine these 2 states into one. We cannot include the key down not allowed keys in this array since repetition keys and not allowed repetition keys are not independent when the key is pressed for the first time as it will get overwritten by whichever of these that come last in the registering of keyboard method.