Implement collision detection inside a grid
The grid here simply represents a number of 128x128 cells where each cell contains 4 32 bit unsigned integers assigned to them. Each frame we iterate through the entities and find which cells(rectangles) do their circle bounds intersect. We then store the index of the entity inside the cell as follows: find one of the appropriate uint amongst those 4 uints inside that cell by dividing the index by 32 and then take modulus by 32 to find the bit place of this uint that we need to set to 1. We then iterate through these cells again and retrieve the indices of the active entities inside the cells and just do collision between these entities for each cell.