Define animation component
The animation component is defined so that if the animation that is requested to be displayed is N frames, then it renders N of the associated textures and it comes to a halt. So it is for a single time use. We also defined a utlity function to be used at main in order to gather all the animations and store their info in the AnimationMetaData vector that gets passed to the engine. We also decided to use vector of std::pair<string, string> instead of unordered_map for storing the associated names of the textures to their locations since we care about the order they are stored in order to know exactly which texture index the animation should start.