Refactor and decouple the components
The components have been rewritten in order to remove unnecessary dependencies and decouple them even more. For example, the ActiveBasedStateComponent and VisibilityBasedStateComponents are made to be the only classes responsible for managing whether an entity is active, or visible respectively. Before that the entity class was responsible for managing the active state which made things confusing as the codebase started to grow. The visibility was managed by graphics component which again should only have one responsibility and that is managing the rendering of the texture associated with our entity not determining if the entity should be visible or not.