Enable GraphicsComponent to handle animation
The graphics component which was meant to be responsible for handling rendering of the owner entity was only able to handle static textures that were not animations. This was limiting as for example we would like our asteroid to spin in space like the real world case. Therefore, it was decided to enable this component to handle graphics as well. However, it does so continuously until the entity is made invisible or inactive. This is in contrast with the AnimationComponent that derives from this and modifies the update method so that it is one time thing for rendering the triggered animation.