Measure active state by being in window bounds
Before this, visibility was one of the criterias of whether something where to be active or not. However, the better criteria is whether it is in window bounds. The reason is that the player should not be deactive when it goes out of bounds of the window. Since not rendering is measured by visibility state, that would cause the player not to be active after going out of bounds which results in it not being able to come back in the scene.