Define Reset method for one time animations
When we go to the next level or would like to repeat the same level again, we need to make sure none of the animations triggered from the level before or the time before we lost get repeated again. Therefore, we define Reset method in order to reset them before starting the new or same level again.