Refactor MemoryPool and allocator code
Exceptions had to be used when a memory pool could not be initialized. The guard value was a bad way to detect memory being overwritten beyond boundary of the type since it could just skip it by some offset. Some of the if checks are unnecessary as a result and had to be removed. The guard value is a good debugging method though and so it was retained.