Fix bug in event management system
The list of delegates corresponding to each event type were being run each time an event was processed from the queue. This meant for example that each collision between each object were processed as many times as there collisions in that specific frame which is definitely not what we intended. In order to fix that we resorted to having one phenomena for each event type in the event queue and have the real events in the vector of delegates. This means that each phenomena of some event can occur once which is why we have an std::unordered_map be an event queue rather than a vector or list.